requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.2
Pillow==9.0.1
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tornado==6.1
typing-extensions==4.1.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.9
